Update cpplint version (#338)
This PR updates our pre-commit to use the latest version of cpplint. This brings in [my fix](https://github.com/cpplint/cpplint/pull/269) that ensures that the pre-commit hook actually works in all environments. Without that fix, depending on the version of Python and setuptools in the base environment the cpplint installation may fail.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)
  - Kyle Edwards (https://github.com/KyleFromNVIDIA)

URL: https://github.com/rapidsai/ucxx/pull/338

+++ cpp/src/remote_key.cpp
@@ -6,6 +6,7 @@
 #include <memory>
 #include <sstream>
 #include <string>
+#include <vector>
 
 #include <ucp/api/ucp.h>
 
@@ -84,7 +85,7 @@ std::shared_ptr<RemoteKey> createRemoteKeyFromSerialized(std::shared_ptr<Endpoin
 
 size_t RemoteKey::getSize() const { return _memorySize; }
 
-uint64_t RemoteKey::getBaseAddress() { return (uint64_t)_memoryBaseAddress; }
+uint64_t RemoteKey::getBaseAddress() { return _memoryBaseAddress; }
 
 ucp_rkey_h RemoteKey::getHandle() { return _remoteKey; }
